backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-migrate==4.0.5
email-validator==2.1.0
python-dotenv==1.0.0